requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
backcall==0.1.0           # via ipython
click==7.0                # via pip-tools
decorator==4.3.2          # via ipython, traitlets
django-debug-toolbar==1.11
django-environ==0.4.5
django==2.1.7
djangorestframework==3.9.1
ipython-genutils==0.2.0   # via traitlets
ipython==7.3.0
jedi==0.13.2              # via ipython
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.3.2
prompt-toolkit==2.0.8     # via ipython
psycopg2==2.7.7
ptyprocess==0.6.0         # via pexpect
pygments==2.3.1           # via ipython
pytz==2018.9              # via django
six==1.12.0               # via pip-tools, prompt-toolkit, traitlets
sqlparse==0.2.4           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
